#90: Adjust groups
